requirements.txt
# Deep Learning Framework
torch==2.1.0
torchvision==0.16.0

# MedMNIST Dataset
medmnist==2.2.3

# BiomedCLIP (via HuggingFace)
transformers==4.35.0
open_clip_torch==2.23.0

# Scientific Computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Progress Bars
tqdm==4.66.1

# Configuration
PyYAML==6.0.1

# Metrics & Evaluation
scikit-learn==1.3.1

# Experiment Tracking (optional)
tensorboard==2.14.1
wandb==0.15.12

# Jupyter
jupyter==1.0.0
ipykernel==6.25.2